icon in notification and some bug fixes

+++ app/src/main/java/com/maxim/musicplayer/audioList/data/AudioData.kt
@@ -7,7 +7,7 @@ data class AudioData(
     private val id: Long,
     private val title: String,
     private val artist: String,
-    private val duration: Int,
+    private val duration: Long,
     private val album: String,
     private val artBitmap: Bitmap?,
     private val uri: Uri
@@ -17,7 +17,7 @@ data class AudioData(
             id: Long,
             title: String,
             artist: String,
-            duration: Int,
+            duration: Long,
             album: String,
             artBitmap: Bitmap?,
             uri: Uri
